% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{addition}
\alias{addition}
\title{Add two numeric values or paste two character values.}
\usage{
addition(a, b)
}
\arguments{
\item{a}{A single value that is numeric, character or that
can be coerced to character.}

\item{b}{A single value that is numeric, character or that
can be coerced to character.}
}
\value{
A numeric value with the sum of a and b, if both
are numeric, or a string if both values can be coerced to
to string.
}
\description{
This function takes two single values and adds them,
if they are numeric, or pastes them together, otherwise.
The function checks the length of each argument, and
returns an error if any of the arguments has length
greater than one.
}
\examples{
# Adding numerc values
addition(1, 6)
# Pasting strings together
addition("nice ", "function")
}
\keyword{addition,}
\keyword{pasting,}
\keyword{sum.}
